# swagger-spec.json
{"openapi":"3.0.0","paths":{"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["Health"]}},"/files/download/{id}":{"get":{"operationId":"FilesController_download","summary":"","description":"Download file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string","format":"binary"}}}},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Files"]}},"/files/show/{id}":{"get":{"operationId":"FilesController_show","summary":"","description":"Show file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string","format":"binary"}}}},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Files"]}},"/files":{"post":{"operationId":"FilesController_create","summary":"Need to be [Admin].","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFileDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Files"],"security":[{"bearer":[]}]},"get":{"operationId":"FilesController_findAll","summary":"Need to be [Admin].","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FileEntity"}}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Files"],"security":[{"bearer":[]}]}},"/files/{id}":{"get":{"operationId":"FilesController_findOne","summary":"Need to be [Admin].","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Files"],"security":[{"bearer":[]}]},"delete":{"operationId":"FilesController_remove","summary":"Need to be [Admin].","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Files"],"security":[{"bearer":[]}]}},"/users/group":{"get":{"operationId":"UsersController_findAllBySameGroupRoute","summary":"Need to be [Admin, Technician]. Retrieves users from the same group as the requesting user.","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, Technician].","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/me":{"get":{"operationId":"UsersController_getMeRoute","summary":"Need to be authentified. Fetches the profile details of the currently authenticated user.","description":"## Access\n### To access this route need to be authentified.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UsersController_updateMeRoute","summary":"Need to be authentified. Updates the profile information of the currently authenticated user.","description":"## Access\n### To access this route need to be authentified.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_removeMeRoute","summary":"Need to be authentified. Deletes the account of the currently authenticated user.","description":"## Access\n### To access this route need to be authentified.","parameters":[],"responses":{"200":{"description":""},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UsersController_findAllRoute","summary":"Need to be [Admin]. Retrieves a list of all users purposes.","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]},"post":{"operationId":"UsersController_createUserAdminRoute","summary":"Need to be [Admin]. Creates a new user with administrative privileges.","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserAdminDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{uuid}":{"get":{"operationId":"UsersController_findOneRoute","summary":"Need to be [Admin]. Finds a specific user by their unique identifier (UUID).","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UsersController_updateRoute","summary":"Need to be [Admin]. Updates the information of a user identified by UUID.","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserWithoutPasswordDto"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_removeRoute","summary":"Need to be [Admin]. Removes a user from the system by their UUID.","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_registerRoute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnAuthDto"}}}},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_loginRoute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnAuthDto"}}}},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Auth"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_requestPasswordResetRoute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"201":{"description":""},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPasswordRoute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""},"503":{"description":"Service Unavailable, maintenance."}},"tags":["Auth"]}}},"info":{"title":"Prunidor API","description":"The API for Prunidor","version":"0.2","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateFileDto":{"type":"object","properties":{"name":{"type":"string"},"path":{"type":"string"},"mimetype":{"type":"string"}},"required":["name","path","mimetype"]},"FileEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"path":{"type":"string"},"mimetype":{"type":"string"},"download_link":{"type":"string"},"show_link":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"version":{"type":"number"}},"required":["id","name","path","mimetype","download_link","show_link","created_at","updated_at","version"]},"ReturnUserWithoutPasswordDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string","minLength":8,"maxLength":30},"group":{"type":"string","nullable":true},"email":{"type":"string"},"role":{"enum":["Admin","Technician","Farmer","User"],"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"version":{"type":"number"}},"required":["id","username","group","email","role","created_at","updated_at","version"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","minLength":8,"maxLength":30}}},"CreateUserAdminDto":{"type":"object","properties":{"password":{"type":"string","minLength":8,"maxLength":30},"email":{"type":"string"},"username":{"type":"string","minLength":8,"maxLength":30},"role":{"enum":["Admin","Technician","Farmer","User"],"type":"string"}},"required":["password","email","username","role"]},"CreateUserDto":{"type":"object","properties":{"password":{"type":"string","minLength":8,"maxLength":30},"email":{"type":"string"},"username":{"type":"string","minLength":8,"maxLength":30}},"required":["password","email","username"]},"ReturnAuthDto":{"type":"object","properties":{"token":{"type":"string"},"id":{"type":"string"},"username":{"type":"string","minLength":8,"maxLength":30},"group":{"type":"string","nullable":true},"email":{"type":"string"},"role":{"enum":["Admin","Technician","Farmer","User"],"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"version":{"type":"number"}},"required":["token","id","username","group","email","role","created_at","updated_at","version"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string","minLength":8,"maxLength":30}},"required":["username","password"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"token":{"type":"string"},"newPassword":{"type":"string","minLength":8,"maxLength":30}},"required":["token","newPassword"]}}}}